pyproject.toml
[project]
name = "django-d3-indicator-viz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-admin-sortable>=2.3",
    "django-admin-sortable2>=2.3",
    "django-filter>=25.2",
    "django-import-export>=4.3.14",
    "django-rest-framework>=0.1.0",
    "factory-boy>=3.3.3",
    "faker>=38.2.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]
